+++ packages/terun-core/src/Generator.ts
@@ -15,6 +15,30 @@ import {
 import * as fs from "fs";
 import { ITransport } from "./types/interfaces";
 
+type TransportParams = {
+  transport: ITransport;
+  source: { [key: string]: any };
+  override?: boolean;
+  debug?: boolean;
+};
+
+type ResolvePathsParams = {
+  transport: ITransport;
+  source: object;
+};
+
+type ResolvePathsResult = { from: string; to: string };
+
+export type DefaultHooks = {
+  global: SyncHook;
+  fileExists: SyncHook;
+  fileSkipped: SyncHook;
+  configure: SyncHook;
+  onTransport: SyncHook;
+  beforeRender: SyncHook;
+  done: SyncHook;
+};
+
 /**
  * Get and render the content files in the
  * destiny file. Use commands in options to create
@@ -25,7 +49,7 @@ class Generator {
   public globalArg: any;
   public render: IRenderEngine;
   public pluginManager: PluginManager;
-  public hooks: { [key: string]: SyncHook };
+  public hooks: DefaultHooks;
 
   constructor(config: IConfigExternal) {
     this.globalConfig = ConfigMapper.fromConfigExternal(config);
@@ -59,13 +83,36 @@ class Generator {
     this.pluginManager.install(this.hooks);
   }
 
+  /**
+   * @typedef
+   *
+   * Resolve from and to properties. It's because the path can have
+   * some properties, for example:
+   *
+   * @example
+   *  const transport = {
+   *    from: "from/path/{{name}}",
+   *    to: "to/path/{{name}}";
+   *  }
+   *
+   *  const source = {
+   *    name: "example"
+   *  }
+   *
+   *  this.resolvePaths({ transport, source })
+   *
+   *  Output: { from: "from/path/example", to: "to/path/example" }
+   *
+   * @param {object} params
+   * @param {string} params.from
+   * @param {string} params.to
+   *
+   * @returns {Promise<{ from: string, to: string }>}
+   */
   public async resolvePaths({
     transport,
     source,
-  }: {
-    transport: ITransport;
-    source: object;
-  }): Promise<{ from: string; to: string }> {
+  }: ResolvePathsParams): Promise<ResolvePathsResult> {
     const { basePath } = this.globalConfig;
     const localSource: object = Object.assign(source);
     const pathFrom: string = path.join(
@@ -76,6 +123,7 @@ class Generator {
         this.globalConfig.tag
       )
     );
+
     const pathTo: string = path.join(
       basePath,
       await this.render.render(transport.to, localSource, this.globalConfig.tag)
@@ -87,17 +135,22 @@ class Generator {
     };
   }
 
+  /**
+   * Transport a item to a file
+   *
+   * @param {Object} param
+   * @param {Object} param.transport
+   * @param {Object} param.source
+   * @param {boolean} param.override
+   * @param {boolean} param.debug
+   * @returns
+   */
   public async transport({
     transport,
     source,
     override = false,
     debug = false,
-  }: {
-    transport: ITransport;
-    source: object;
-    override?: boolean;
-    debug?: boolean;
-  }): Promise<void> {
+  }: TransportParams): Promise<void> {
     this.hooks.configure.call(this.globalConfig);
     this.hooks.onTransport.call(transport, source);
 
@@ -115,7 +168,7 @@ class Generator {
       }
     }
 
-    const localSourcePlugin: any =
+    const localSourcePlugin: { [key: string]: any } =
       (await this.hooks.beforeRender.promise(
         localSource,
         transport,
@@ -127,7 +180,7 @@ class Generator {
     /**
      * The file need exists and command override is different that true
      */
-    const fileExists = fs.existsSync(resolvedPaths.to);
+    const fileExists = Utils.File.existFile(resolvedPaths.to);
     if (fileExists && override) {
       const overrideChoice = await this.hooks.fileExists.promise();
       if (!overrideChoice) {
@@ -147,7 +200,6 @@ class Generator {
     );
 
     Utils.File.createDir(resolvedPaths.to);
-
     Utils.File.writeUtf8File(resolvedPaths.to, fromContentRendered);
 
     this.hooks.done.call(transport, source);