+++ packages/terun-core/src/index.ts
@@ -1,4 +1,5 @@
 import Generator from "./Generator";
 import Render from "./types/render/RenderEngineFactory";
-import Utils from './utils/index'
+import Utils from "./utils/index";
+
 export { Generator, Render, Utils };