+++ CHANGELOG.md
@@ -3,6 +3,27 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.5](https://github.com/raphaelkieling/terun/compare/v1.2.4...v1.2.5) (2021-05-06)
+
+**Note:** Version bump only for package terun
+
+
+
+
+
+## [1.2.4](https://github.com/raphaelkieling/terun/compare/v1.2.3...v1.2.4) (2021-05-06)
+
+
+
+## [1.2.2](https://github.com/raphaelkieling/terun/compare/v1.2.1...v1.2.2) (2021-03-26)
+
+**Note:** Version bump only for package terun
+
+
+
+
+
+
 ## [1.2.3](https://github.com/raphaelkieling/terun/compare/v1.2.1...v1.2.3) (2021-03-26)